requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Caching==2.0.2
Flask-Session==0.4.0
SQLAlchemy==2.0.4
pymysql==1.0.3
pandas
waitress==2.1.2
jinja2==3.1.2
markupsafe==2.1.2
werkzeug==2.2.3
python-dotenv==1.0.0
python-docx>=0.8.11
bcrypt==4.0.1
loguru==0.7.2
urllib3==2.0.7
requests==2.32.3
python-dateutil==2.8.2
flask_migrate
flask_cors
flask_login==0.6.3
cryptography
pywin32>=306; sys_platform == 'win32'